Transport: Refactor transports to use transport template to avoid generating multiple templates per entry
