+++ src/game/Maps/TransportMgr.h
@@ -21,6 +21,7 @@
 
 #include "Platform/Define.h"
 #include "Server/DBCStructure.h"
+#include "Movement/spline.h"
 
 #include <map>
 
@@ -42,6 +43,55 @@ struct TransportAnimation
 
 typedef std::map<uint32, TransportAnimation> TransportAnimationContainer;
 
+typedef Movement::Spline<double>                 TransportSpline;
+
+struct KeyFrame
+{
+    explicit KeyFrame(TaxiPathNodeEntry const& _node) : Index(0), Node(&_node), InitialOrientation(0.0f),
+        DistSinceStop(-1.0f), DistUntilStop(-1.0f), DistFromPrev(-1.0f), TimeFrom(0.0f), TimeTo(0.0f),
+        Teleport(false), Update(false), ArriveTime(0), DepartureTime(0), Spline(nullptr), NextDistFromPrev(0.0f), NextArriveTime(0)
+    {
+    }
+
+    uint32 Index;
+    TaxiPathNodeEntry const* Node;
+    float InitialOrientation;
+    float DistSinceStop;
+    float DistUntilStop;
+    float DistFromPrev;
+    float TimeFrom;
+    float TimeTo;
+    bool Teleport;
+    bool Update;
+    uint32 ArriveTime;
+    uint32 DepartureTime;
+    TransportSpline* Spline;
+
+    // Data needed for next frame
+    float NextDistFromPrev;
+    uint32 NextArriveTime;
+
+    bool IsTeleportFrame() const { return Teleport; }
+    bool IsUpdateFrame() const { return Update; }
+    bool IsStopFrame() const { return Node->actionFlag == 2; }
+};
+
+typedef std::vector<KeyFrame>  KeyFrameVec;
+
+struct TransportTemplate
+{
+    TransportTemplate() : inInstance(false), pathTime(0), accelTime(0.0f), accelDist(0.0f), entry(0) { }
+    ~TransportTemplate();
+
+    std::set<uint32> mapsUsed;
+    bool inInstance;
+    uint32 pathTime;
+    KeyFrameVec keyFrames;
+    float accelTime;
+    float accelDist;
+    uint32 entry;
+};
+
 class TransportMgr
 {
     public:
@@ -50,17 +100,23 @@ class TransportMgr
 
         TransportAnimation const* GetTransportAnimInfo(uint32 entry) const
         {
-            TransportAnimationContainer::const_iterator itr = m_transportAnimations.find(entry);
+            auto itr = m_transportAnimations.find(entry);
             if (itr != m_transportAnimations.end())
                 return &itr->second;
 
             return nullptr;
         }
 
+        TransportTemplate const* GetTransportTemplate(uint32 entry);
+
+        void LoadTransportTemplates();
+
     private:
         void AddPathNodeToTransport(uint32 transportEntry, uint32 timeSeg, TransportAnimationEntry const* node);
+        bool GenerateWaypoints(GameObjectInfo const* goinfo, TransportTemplate& transportTemplate);
 
         TransportAnimationContainer m_transportAnimations;
+        std::unordered_map<uint32, TransportTemplate> m_transportTemplates;
 };
 
 #define sTransportMgr MaNGOS::Singleton<TransportMgr>::Instance()
